feat: restructure file
move components and global.css to src/ with corresponding folder
