+++ components.json
@@ -5,7 +5,7 @@
   "tsx": true,
   "tailwind": {
     "config": "",
-    "css": "src/app/globals.css",
+    "css": "src/style/globals.css",
     "baseColor": "zinc",
     "cssVariables": true,
     "prefix": ""
