+++ src/components/landing/Featured-products.tsx
